refactor: Extract Tag unused style (#44512)
* refactor: tag status as extract style

* chore: add key

* chore: preset of it
